requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.3.1
frozenlist==1.4.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
mozia-modules==1.8.3
multidict==6.0.4
PyMySQL==1.1.0
python-dotenv==1.0.0
redis==3.5.3
redis-py-cluster==2.1.3
requests==2.31.0
soupsieve==2.5
urllib3==2.0.7
yarl==1.9.2
